requirements.txt
pillow
transformers
yolov5
pytesseract
opencv-python
numpy
torch
sympy